`TwoWayEnumerable` 簡略化＆`SearchTextParser` の不具合を修正
